Refactor audio controls
